+++ src/shared/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './hooks';
 export * from './locales/i18nConfiguration';
